pyproject.toml
[tool.poetry]
name = "weather-forecast"
version = "0.1.0"
description = ""
authors = ["Igor <igorbilykwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
aiohttp = "^3.10.8"
aioboto3 = "^13.1.1"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
